Update Header component
